Terminando o Sistema de RH

+++ trabalhoPratico1-Peterson/remover_departamento.php
@@ -1,8 +1,20 @@
 <?php
-    include "conexao.php";
+    // Aqui eu verifico se o usuario esta logado;
+    require_once "verificar_usuario.php";
 
-    $id = $_POST['id'];
-    $sql = "DELETE FROM DEPARTAMENTO WHERE ID_DEPARTAMENTO = '$id'";
-    $conexao->query($sql);
-    echo"1";
+    // Confiro se o usuario tem a permissao de remover dados, se tiver ele consegue se nao ele é redirecionado
+    if($_SESSION["perfil"] != "1"){
+        header("location:index.php");
+    }
+    else{
+        require_once "conexao.php";
+
+        // Aqui eu recebo e removo os dados
+        $id = $_POST['id'];
+        $sql = "DELETE FROM DEPARTAMENTO WHERE ID_DEPARTAMENTO = '$id'";
+        $conexao->query($sql);
+
+        // Aqui eu retorno "1" p/ realizar outras funcoes no sistema, como exibir uma msg de sucesso
+        echo"1";
+    }
 ?>
