+++ trabalhoPratico1-Peterson/remover_funcao.php
@@ -1,9 +1,20 @@
 <?php
-    include "conexao.php";
+    // Aqui eu verifico se o usuario esta logado;
+    require_once "verificar_usuario.php";
 
-    $id = $_POST['id'];
-    $sql = "DELETE FROM FUNCAO WHERE ID_FUNCAO = '$id'";
-    $conexao->query($sql);
+    // Confiro se o usuario tem a permissao de remover dados, se tiver ele consegue se nao ele é redirecionado
+    if($_SESSION["perfil"] == "3"){
+        header("location:index.php");
+    }
+    else{
+        require_once "conexao.php";
 
-    echo"1";
+        // Aqui eu recebo e removo os dados
+        $id = $_POST['id'];
+        $sql = "DELETE FROM FUNCAO WHERE ID_FUNCAO = '$id'";
+        $conexao->query($sql);
+
+        // Aqui eu retorno "1" p/ realizar outras funcoes no sistema, como exibir uma msg de sucesso
+        echo"1";
+    }
 ?>
